Rearranging the structure

+++ Ex.1-2 (Data types, Variables)/Ex.2-4 Tips calculator.py
@@ -11,11 +11,11 @@ bill = input("What was the total bill? $")
 tip = input("What percentage tip would you like to give? 10, 12, or 15? ")
 people = input("How many people to split the bill? ")
 percentage = (int(tip)/100)+1
-result = round((int(bill)/int(people))*percentage, 2)
+result = round((float(bill)/int(people))*percentage, 2)
 print(f"Each person should pay: ${result}")
 
 # Not mine:
-# print("Welcom to the tip calculator!")
+# print("Welcome to the tip calculator!")
 # bill = float(input("What was the total bill? $"))
 # tip = int(input("What percentage tip would you like to give? 10, 12, or 15? "))
 # people = int(input("How many people to split the bill? "))
